requirements.txt
Django>=4.1.7,<5.0
pyyaml>=6.0,<7.0
djangorestframework>=3.14.0,<4.0
djangorestframework-simplejwt>=5.2.2,<6.0
pytest-django>=4.5.2,<5.0